+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added a `Delay` structure that implements the embedded-hal delay traits.
+
 ## [0.3.0] - 2021-02-14
 ### Changed
 - Improved latency for GPIOs
@@ -23,8 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2020-09-12
 - Initial release
 
-[Unreleased]: https://github.com/newAM/ftd2xx-embedded-hal/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/newAM/ftd2xx-embedded-hal/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/newAM/ftd2xx-embedded-hal/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/newAM/ftd2xx-embedded-hal/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/newAM/ftd2xx-embedded-hal/releases/tag/v0.1.0
-